Add ignore file and ignore hardware config
The nixos hardware configuration file is fairly specific to your own
system and is generated by a nixos-rebuild if it doesn’t already
exist.

If it were included in the repository it would make it trickier to use
this config on another system and would make hacking on the hardware
of any system more burdensome as you would have some incentive to
track the changes along with your environment changes.