Add a the user zennad
* Add zennad
* Add an inital password
* Have the user be immutable, but add the comment for it to be
adjusted to mutability
* Add the user to the wheel and sudo groups

As the user is immutable by default the inital password acts as the
only password.

I’m uncertain whether I want the user to be a normal user or a system
user.

For the moment I have it set up as a normal user but this may change.