Describe issues with flakes and configuration.nix
I am trying to use Nix flakes with my configuration.nix
however there is an issue with the config file importing the hardware
config.
This makes sense as the import is simply a local path and so there is
no guarantee that the hardware config file has not changed, breaking
reproducability.

I think that I can include the hardware config in the flake, but I
can’t seem to get my config file to be able to find the hardware
config to be able to import it.

I don’t know if it’s just that I don’t know what I’m doing (highly
probable), or if it’s because flakes are still experimental, or if you
simply cannot do what I want to do and I should try another method.

Hmmm....
more experimenting and research is required