Add comments with ideas for including my hardware config in a flake
I want to have my nixos config file become independant of my
hardware-configuration.

My idea of how to do this is to have my hardware config be passed in
as an argument to my config so it can add it to the imports list.
If this works I’ll be able to use the config file on other systems
with a completely different hardware config.

I’ll have my config in a separeate git repo, probably hosted on github
or gitlab. I’ll then have the flake on my computer import that config
flake from the repo and have it use a system local hardware config
file.

Hopefully along the way I’ll be able to find a solution to my issue of
not knowing a good substitute for impure nix paths.
Using flake purely, you can’t write this:
“
import <nixpkgs/nixos/modules/installer/scan/not-detected.nix> {}
”
This isn’t allowed as it refers to the impure NIX_PATH env variable
which can be modified without control from nix; and infact can be
changed within nix derivations as well.

With my current understanding of how flakes work, I think that I can
specify the module in the flake file and then pass it in as an input
to the hardware file.

“
...
  inputs = {
    nixpkgs.url = "github:NixOS/nixpkgs/nixos-20.03";
...
  outputs = {
    ?.nixossystem = import ./systems.nix nixpkgs.nixosModules;
...
”

We’ll see if I’m on the right track with this idea.