Add *.lock files to gitignore
Make git ingore lock files; generated by nix flakes in particular but
it might be alright to ignore all ‘.lock’ files

Tbh, I was unsure if this is the best practice for flakes
You can, at least in theory, specify all the information you might
need to get a very specific version of a repo that you are using as
input to a flake, so tracking the lock file seems only useful as a
means of historical documentation of which version were used during a
period of time. Maybe I’ve missed something here.