On the organisation of my nixos config
I think I will try to abstract out the hardware config from this repo
and make a separeate repo with just the specific hardware config for
any machine which I use nixos on.
I would like to have my main config in a flake repo and just make a
hardware config flake on the machine, add my main config to the inputs
and include it as a module, then have a way to fairly simply share my
desired config across machines with any desired changes being added
machine to machine.