Remove user specific configuration
This will remove configuration for my user account and
the password for root.

This is with the goal of generalising my configuration
to be reusable on perhaps a different machine where I might
not wish for there to be this specific account.

I will then make another module with any user configuration that
I need and merge this module with it.